Begin SaaH
TODO: Find out what causes the server to halt on the packet decoder causing it
to miss acknowledging packets.
TODO: Add a better way of detecting the client setting their own name/colour
and updating gamedata in the process (which it is not the owner of). See
ln.193 WorkerNode.ts. Some solutions include:
 - Check if the gamedata data is 'normal', i.e. it only shows their updated
name/colour/hat/skin/pet which might be annoying but could work better than
solution 2.
 - Wait for checkname/checkcolour before telling the client that they are an
acting host. Probably doesn't work for set hat/skin/pet, but I'm not sure if
that even has a GameData update packet with it (?)"